feat(molecule/photoUploader): create PhotosPreview component and extract its funcs
